+++ test/input_test.rb
@@ -2,19 +2,19 @@ require_relative "test_helper"
 
 class InputTest < Minitest::Test
   def test_input_format_stderr
-    assert_index_file "queries.log", "stderr"
+    assert_index_file "queries.log"
   end
 
   def test_input_format_csv
-    assert_index_file "queries.csv", "csv"
+    assert_index_file "queries.csv"
   end
 
   def test_input_format_json
-    assert_index_file "queries.json", "json"
+    assert_index_file "queries.json"
   end
 
   def test_input_format_sql
-    assert_index_file "queries.sql", "sql"
+    assert_index_file "queries.sql"
   end
 
   def test_pg_stat_activity
@@ -102,9 +102,9 @@ class InputTest < Minitest::Test
 
   private
 
-  def assert_index_file(file, input_format)
+  def assert_index_file(file)
     file = File.expand_path("../support/#{file}", __FILE__)
-    assert_dexter_output "Index found: public.posts (id)", [file, "--input-format", input_format]
+    assert_dexter_output "Index found: public.posts (id)", [file]
   end
 
   def setup_pg_stat_monitor